requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

botasaurus==4.0.97
ImageHash==4.3.2
pandas==3.0.1
pillow==12.1.1
playwright==1.58.0
PyQt6==6.10.2
python-dotenv==1.2.1
requests==2.32.5
tqdm==4.67.3
xxhash==3.6.0
snakemake==9.16.3
pyarrow==23.0.1

